Clarified the terminology and fixed some bugs
Used attributes consistently to refer to members of states and
data-members for other uses. The term _owners_ is now used instead of
signatories.

Notary now adds a timestamp, computes a transaction_id by hashing the
transaction contents, and signs to the transaction. The transaction_id
is returned to the caller.

Clarified how the list of owners is computed.

Clarified the difference between public and private inputs to the ZKP
program